data type changes and ER diagram

+++ ATM_Database/storedProcedures/update_Balance.sql
@@ -1,6 +1,6 @@
 DELIMITER $$
 
-CREATE PROCEDURE update_Balance(IN inPAN BIGINT, IN updatedBalance BIGINT)
+CREATE PROCEDURE update_Balance(IN inPAN BIGINT, IN updatedBalance DECIMAL(10, 2))
 BEGIN
     update userAccount
     set Balance = updatedBalance
